# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  httpx.JSendError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Failed to connect to database
        type: string
      status:
        example: error
        type: string
    type: object
  httpx.JSendFail:
    properties:
      data:
        additionalProperties: {}
        type: object
      status:
        example: fail
        type: string
    type: object
...
  httpx.JSendSuccess:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  internal_notifications_models.PaginationMetadata:
    properties:
      current_page:
        example: 1
        type: integer
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      next_url:
        example: /api/v1/notifications?page=2&limit=20
        type: string
      per_page:
        example: 20
        type: integer
      previous_url:
        type: string
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  models.DeviceType:
    enum:
    - android
    - ios
    - web
    type: string
    x-enum-varnames:
    - DeviceTypeAndroid
    - DeviceTypeIOS
    - DeviceTypeWeb
...
  models.FCMToken:
    properties:
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      device_id:
        example: device-abc-123
        type: string
      device_type:
        allOf:
        - $ref: '#/definitions/models.DeviceType'
        enum:
        - android
        - ios
        - web
        example: android
      id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
      is_active:
        example: true
        type: boolean
      last_used_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      token:
        example: fL8X9Y2Z3A4B5C6D7E8F9G0H1I2J3K4L5M6N7O8P9Q0R1S2T3U4V5W6X7Y8Z9A0B1C
        type: string
      updated_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      user_id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
    type: object
...
  models.Notification:
    properties:
      body:
        example: 'Tu pedido #1234 está en camino'
        type: string
      created_at:
        example: "2025-01-15T10:00:00Z"
        type: string
      data:
        type: string
      id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
      is_read:
        example: false
        type: boolean
      notification_type:
        allOf:
        - $ref: '#/definitions/models.NotificationType'
        enum:
        - order_created
        - order_updated
        - order_assigned
        - order_in_transit
        - order_delivered
        - order_canceled
        - payment_received
        - payment_failed
        - driver_assigned
        - driver_nearby
        - general
        - promotional
        example: order_in_transit
      read_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      title:
        example: Pedido en camino
        type: string
      user_id:
        example: d53b655c-e833-400e-b0e8-ee68ea18e2cc
        type: string
    type: object
  models.NotificationListResponse:
    properties:
      data:
        properties:
          items:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
          pagination:
            $ref: '#/definitions/internal_notifications_models.PaginationMetadata'
        type: object
      status:
        example: success
        type: string
    type: object
  models.NotificationResponse:
    properties:
      data:
        $ref: '#/definitions/models.Notification'
      status:
        example: success
        type: string
    type: object
  models.NotificationType:
    enum:
    - order_created
    - order_updated
    - order_assigned
    - order_in_transit
    - order_delivered
    - order_canceled
    - payment_received
    - payment_failed
    - driver_assigned
    - driver_nearby
    - general
    - promotional
    type: string
    x-enum-varnames:
    - NotificationTypeOrderCreated
    - NotificationTypeOrderUpdated
    - NotificationTypeOrderAssigned
    - NotificationTypeOrderInTransit
    - NotificationTypeOrderDelivered
    - NotificationTypeOrderCanceled
    - NotificationTypePaymentReceived
    - NotificationTypePaymentFailed
    - NotificationTypeDriverAssigned
    - NotificationTypeDriverNearby
    - NotificationTypeGeneral
    - NotificationTypePromotional
...
  models.RegisterTokenRequest:
    properties:
      device_id:
        example: device-abc-123
        type: string
      device_type:
        allOf:
        - $ref: '#/definitions/models.DeviceType'
        enum:
        - android
        - ios
        - web
        example: android
      token:
        example: fL8X9Y2Z3A4B5C6D7E8F9G0H1I2J3K4L5M6N7O8P9Q0R1S2T3U4V5W6X7Y8Z9A0B1C
        type: string
    required:
    - device_type
    - token
    type: object
  models.TokenResponse:
    properties:
      data:
        $ref: '#/definitions/models.FCMToken'
      status:
        example: success
        type: string
    type: object
  models.UnreadCountResponse:
    properties:
      data:
        properties:
          count:
            example: 5
            type: integer
        type: object
      status:
        example: success
        type: string
    type: object
  models.UnregisterTokenRequest:
    properties:
      token:
        example: fL8X9Y2Z3A4B5C6D7E8F9G0H1I2J3K4L5M6N7O8P9Q0R1S2T3U4V5W6X7Y8Z9A0B1C
        type: string
    required:
    - token
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Delivery marketplace API for customers, merchants, and drivers.
  title: TacoShare Delivery API
  version: "1.0"
paths:
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Get paginated list of notifications for the authenticated user.
        Supports pagination with page and limit query parameters. Default page size
        is 20, maximum is 100.
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 'Items per page (default: 20, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved notifications
          schema:
            $ref: '#/definitions/models.NotificationListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error - failed to retrieve notifications
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific notification. User must own the notification
        to delete it. This permanently removes the notification from the database.
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification deleted successfully
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "400":
          description: Invalid notification ID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "403":
          description: Forbidden - notification belongs to another user
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error - failed to delete notification
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Delete notification
      tags:
      - notifications
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific notification. User must
        own the notification to access it.
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved notification
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "400":
          description: Invalid notification ID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "403":
          description: Forbidden - notification belongs to another user
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error - failed to retrieve notification
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get notification details
      tags:
      - notifications
  /notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a specific notification as read. This updates the is_read
        field and sets the read_at timestamp. User must own the notification.
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Notification marked as read successfully
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "400":
          description: Invalid notification ID
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "403":
          description: Forbidden - notification belongs to another user
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error - failed to update notification
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /notifications/read-all:
    put:
      consumes:
      - application/json
      description: Mark all unread notifications for the authenticated user as read.
        This updates all notifications with is_read=false to is_read=true and sets
        their read_at timestamps.
      produces:
      - application/json
      responses:
        "200":
          description: All notifications marked as read successfully
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error - failed to update notifications
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/register-token:
    post:
      consumes:
      - application/json
      description: Register a Firebase Cloud Messaging token for push notifications.
        Tokens are device-specific and allow sending notifications to that device.
        If the token already exists, it will be updated as active.
      parameters:
      - description: FCM token details (token, device_type, optional device_id)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token registered successfully
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error - failed to register token
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Register FCM token
      tags:
      - notifications
...
  /notifications/unread-count:
    get:
      consumes:
      - application/json
      description: Get the count of unread notifications for the authenticated user.
        Useful for displaying notification badges.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved unread count
          schema:
            $ref: '#/definitions/models.UnreadCountResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "500":
          description: Internal server error - failed to count notifications
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - notifications
  /notifications/unregister-token:
    delete:
      consumes:
      - application/json
      description: Deactivate a Firebase Cloud Messaging token. Use this when the
        user logs out or uninstalls the app to stop receiving notifications on that
        device.
      parameters:
      - description: FCM token to deactivate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UnregisterTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token unregistered successfully
          schema:
            $ref: '#/definitions/httpx.JSendSuccess'
        "400":
          description: Invalid request body or missing token
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/httpx.JSendError'
        "404":
          description: Token not found
          schema:
            $ref: '#/definitions/httpx.JSendFail'
        "500":
          description: Internal server error - failed to unregister token
          schema:
            $ref: '#/definitions/httpx.JSendError'
      security:
      - BearerAuth: []
      summary: Unregister FCM token
      tags:
      - notifications
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"